agentic_rag/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.11
langgraph>=0.0.19
python-dotenv>=1.0.0
neo4j>=5.15.0
tavily-python>=0.2.6
numpy>=1.22.0
sentence-transformers>=2.2.2